requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==4.4.2
fastapi==0.89.1
gcloud==0.18.3
googleapis-common-protos==1.58.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
jws==0.1.3
moviepy==1.0.3
numpy==1.24.1
oauth2client==4.1.3
Pillow==9.4.0
proglog==0.1.10
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.17
pydantic==1.10.4
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==2.0.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tqdm==4.64.1
typing_extensions==4.4.0
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
